front: drop importTimetableItemConfig
